dom/notification/test/mochitest/test_bug931307.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<html>
<head>
  <title>Bug 931307</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<pre id="test">
<script type="application/javascript"><!--

SimpleTest.waitForExplicitFinish();
var notification = new Notification("");
var promise = Notification.get();
promise.then(
  function onSuccess() {
    ok(true, "No error when creating a notification without title");
  },
  function onFailure() {
    ok(false, "Should not throw error when creating a notification without title");
  }
).then(() => {
  notification.close();
}).then(() => {
  SimpleTest.finish();
});

</script>
</pre>
</body>
</html>
